Create CountPrimes
The code I wrote in the exercise with them yesterday was UNBELIEVABLY non-characteristic of me. Generally I avoid using break as much as I can and that's the first thing I went for. The looping was convoluted and the inner loop was counting backwards for no good reason. The code for IsPrime method was spread within the body of ListOfPrimes like a splatter of diarrhea in a toilet bowl.

I wrote the above code in 9 minutes.